src/requirements.txt
numpy>=1.21.2
pandas>=1.3.3
matplotlib>=3.4.3
scikit-learn>=0.24.2
torch>=1.9.0
jason>=1.3.0
threading>=3.0.10
plotly.graph_objects>=5.3.0
scipy>=1.7.1
networkx>=2.6.3
ast>=0.0.1
xml>=1.0.0
tqdm>=4.62.3
